shared/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
networkx==3.2.1
google-generativeai==0.4.0
elevenlabs==1.0.0
pydantic==2.6.1
python-dotenv==1.0.1
